backend/requirements.txt
# AegisCX Backend Dependencies
# Python 3.11+ required
# Install: pip install -r requirements.txt

# ─── Web Framework ─────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9          # File upload support
httpx==0.27.0                    # Async HTTP client

# ─── Database ──────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0                  # Async PostgreSQL driver
aiosqlite==0.20.0                # Async SQLite driver (local dev fallback)
alembic==1.13.1                  # Database migrations

# ─── Task Queue ────────────────────────────────────────────
celery[redis]==5.3.6
redis==5.0.4
flower==2.0.1                    # Celery monitoring

# ─── Authentication & Security ─────────────────────────────
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4            # Password hashing
slowapi==0.1.9                    # Rate limiting

# ─── Configuration ─────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1

# ─── Audio Processing ──────────────────────────────────────
ffmpeg-python==0.2.0
noisereduce==3.0.2
soundfile==0.12.1
librosa==0.10.2.post1
webrtcvad-wheels                 # Better Windows compatibility than raw webrtcvad
numpy
scipy

# ─── Speech-to-Text ────────────────────────────────────────
# WhisperX for production (install separately with GPU support)
# whisperx==3.1.5               # pip install git+https://github.com/m-bain/whisperX.git
faster-whisper>=1.0.3         # Loosened to resolve tokenizer conflicts

# ─── NLP + ML ──────────────────────────────────────────────
transformers>=4.40.0
torch                           # CPU version; for GPU: torch==2.3.0+cu121
torchaudio
sentence-transformers==3.0.0
scikit-learn==1.5.0
bertopic==0.16.1
umap-learn==0.5.6

# ─── LLM APIs ──────────────────────────────────────────────
google-generativeai==0.7.0      # Gemini Pro
openai==1.30.1                  # GPT-4o fallback

# ─── Vector Database ───────────────────────────────────────
chromadb==0.5.0

# ─── Report Generation ─────────────────────────────────────
weasyprint==62.3                # PDF generation
jinja2==3.1.4                   # HTML templates for reports

# ─── Utilities ─────────────────────────────────────────────
structlog==24.1.0               # Structured logging
tenacity==8.3.0                 # Retry logic with backoff
python-dateutil==2.9.0
pytz==2024.1

# ─── Testing ───────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0                   # Test client for FastAPI
factory-boy==3.3.0              # Test data factories
mistralai                       # Mistral API
